aws-data-warehouse-modeling/requirements.txt
psycopg2-binary
pandas
numpy
jupyter
ipython-sql
keyring
boto3